Déplace les routes dans le dossier `routes`
