+++ scripts/front-end/config.js
@@ -1,0 +1,4 @@
+//@ts-check
+
+/** @type {Element | null} */
+export const svelteTarget = document.querySelector('.svelte-main')
