+++ scripts/front-end/mapStateToComponentProps.js
@@ -1,0 +1,7 @@
+import store from './store.js'
+
+export const mapStateToSqueletteProps = (state) => {
+    return {
+        email: store.state.secret ? '@' : undefined
+    }
+}
